Issue(#918) - Heatmap now updates on activity
Previously, the timeline was implemented through a hook `useTimeline`
that on page load would create a request object and fetch the
current activity status for the heatmap (and other timeline users).

However, when activity was updated, there was no mechanism to signal
the `useTimeline` hook to update its information, since each instance
of the hook was local the component it was mounted with.

`useTimelime` was therefore changed to `useTimelineContext`, which
allows the updateUserProgress call to signal the (now global)
timeline context to update the locally stored activity.