fix(clock): round ideal time usage to nearest second (#1678)
* Clock Usage: Round expectedTime to nearest second

When dealing with the ideal time, fractions of a
second shouldn't be treated as significant digits.

This change rounds down any decimals when
calculating that ideal time.

* Reported as part of Issue #1506

* clockRating: Use round instead of floor

Final results can be any number of seconds so
don't need to avoid the value of 60 seconds like
we do when formatting time for human-use.

---------

Co-authored-by: Alex Gustafson <alex@weareadjacent.com>