feat: use GameTable in the import game form (#1704)
* feat: use GameTable in the import game form

This allows the set of games to be filtered and sorted with a similar
interface to that used in other places (e.g., the My Games page).

* feat: display rating provisionality on GameTable

This adds the field to GameInfo and the rendering, but N.B. that there
isn't always the setup to keep this information carried around
everywhere (e.g., the mark displays on OnlineGameForm, but the
information is currently lost during import to the My Games page).

* fix: more control of free tier restrictions in pagination

CustomPagination checked free tier status and unconditionally limited
access to the first page depending on that. Since that's not always
appropriate, expose a parameter that allows customizing whether free
tier status should matter. GameTable (which uses CustomPagination) already
had such a parameter, so we pass that along to CustomPagination.

* Revert "feat: display rating provisionality on GameTable"

This reverts commit d4337b5663ebde924a3da7f005c1401b109df7d4.

This way of doing it played pretty fast and loose with types,
so on second thought let's not do that.

* cleanup cleanup cleanup